+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/common/block/FlexibleStatue.kt
@@ -106,4 +106,13 @@ class FlexibleStatue :
             context.horizontalDirection.opposite,
         )
     }
+
+    override fun getCloneItemStack(blockGetter: BlockGetter, blockPos: BlockPos, blockState: BlockState): ItemStack {
+        val blockEntity = blockGetter.getBlockEntity(blockPos)
+        if (blockEntity is FlexibleStatueBlockEntity) {
+            return prepareItemStack(blockEntity, blockState)
+        }
+        @Suppress("DEPRECATION")
+        return super.getCloneItemStack(blockGetter, blockPos, blockState)
+    }
 }